feat: refactor DashboardPage structure and add section components for better organization
